requirements.txt
django
psycopg2-binary
dj-database-url
redis
django-filter
whitenoise
celery
celery[redis]
pillow
gunicorn
django-cloudinary-storage
cloudinary
django-guardian
django-bootstrap5
django-allauth
django-anymail[mailjet]